@@ -1,27 +1,25 @@
 import { useState } from "react";
 
+import { ptBR } from "date-fns/locale/pt-BR";
 import MUIDataTable from "mui-datatables";
+import { AdapterDateFns } from "@mui/x-date-pickers/AdapterDateFnsV3";
+import { DatePicker } from "@mui/x-date-pickers/DatePicker";
+import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
 
-import DateFnsUtils from "@date-io/date-fns";
-import {
-  MuiPickersUtilsProvider,
-  KeyboardDatePicker,
-} from "@material-ui/pickers";
+import Button from "@mui/material/Button";
+import CircularProgress from "@mui/material/CircularProgress";
+import FormGroup from "@mui/material/FormGroup";
+import FormLabel from "@mui/material/FormLabel";
+import Dialog from "@mui/material/Dialog";
+import DialogActions from "@mui/material/DialogActions";
+import DialogContent from "@mui/material/DialogContent";
+import DialogTitle from "@mui/material/DialogTitle";
 
-import Button from "@material-ui/core/Button";
-import CircularProgress from "@material-ui/core/CircularProgress";
-import FormGroup from "@material-ui/core/FormGroup";
-import FormLabel from "@material-ui/core/FormLabel";
-import Dialog from "@material-ui/core/Dialog";
-import DialogActions from "@material-ui/core/DialogActions";
-import DialogContent from "@material-ui/core/DialogContent";
-import DialogTitle from "@material-ui/core/DialogTitle";
-
-import AddIcon from "@material-ui/icons/Add";
-import DeleteIcon from "@material-ui/icons/Delete";
-import EditIcon from "@material-ui/icons/Edit";
-import IconButton from "@material-ui/core/IconButton";
-import Container from "@material-ui/core/Container";
+import AddIcon from "@mui/icons-material/Add";
+import DeleteIcon from "@mui/icons-material/Delete";
+import EditIcon from "@mui/icons-material/Edit";
+import IconButton from "@mui/material/IconButton";
+import Container from "@mui/material/Container";
 
 import { ExpenseForm } from "../../forms/ExpenseForm";
 import { FormFeedback } from "../FormFeedback";
@@ -33,8 +31,8 @@ import {
   ExpensesSourcesMapping,
 } from "../../consts.js";
 
-import PlusOneIcon from "@material-ui/icons/PlusOne";
-import Tooltip from "@material-ui/core/Tooltip";
+import PlusOneIcon from "@mui/icons-material/PlusOne";
+import Tooltip from "@mui/material/Tooltip";
 
 const ExpenseCreateEditDialog = ({ data, open, onClose, reloadTable }) => {
   return (
@@ -122,7 +120,7 @@ export const ExpensesTable = () => {
   const [pageSize, setPageSize] = useState(5);
   const [startDate, setStartDate] = useState(null);
   const [endDate, setEndDate] = useState(
-    new Date(now.getFullYear(), now.getMonth() + 1, 0)
+    new Date(now.getFullYear(), now.getMonth() + 1, 0),
   );
   const [filters, setFilters] = useState({
     page: 1,
@@ -157,7 +155,7 @@ export const ExpensesTable = () => {
     });
 
     Object.entries(multipleChoiceFilters).forEach(([key, value]) =>
-      value.map((v) => _filters.append(key, v))
+      value.map((v) => _filters.append(key, v)),
     );
 
     return _filters.toString();
@@ -244,7 +242,7 @@ export const ExpensesTable = () => {
           : [...ExpensesCategoriesMapping, ...ExpensesSourcesMapping];
 
       let _filters = filterList[changedColumnIndex].map(
-        (f) => getChoiceByLabel(f, mapping).value
+        (f) => getChoiceByLabel(f, mapping).value,
       );
       setFilters({ ...filters, [column]: _filters, page: 1 });
     },
@@ -329,10 +327,13 @@ export const ExpensesTable = () => {
         filterOptions: {
           names: [],
           display: (filterList, onChange, index, column) => (
-            <MuiPickersUtilsProvider utils={DateFnsUtils}>
+            <LocalizationProvider
+              dateAdapter={AdapterDateFns}
+              adapterLocale={ptBR}
+            >
               <FormLabel>Quando</FormLabel>
               <FormGroup row>
-                <KeyboardDatePicker
+                <DatePicker
                   disableToolbar
                   variant="inline"
                   format="dd/MM/yyyy"
@@ -348,7 +349,7 @@ export const ExpensesTable = () => {
                   }}
                   style={{ width: "48%", marginRight: "2%" }}
                 />
-                <KeyboardDatePicker
+                <DatePicker
                   disableToolbar
                   variant="inline"
                   format="dd/MM/yyyy"
@@ -365,7 +366,7 @@ export const ExpensesTable = () => {
                   style={{ width: "48%" }}
                 />
               </FormGroup>
-            </MuiPickersUtilsProvider>
+            </LocalizationProvider>
           ),
         },
       },
