@@ -1,28 +1,26 @@
 import { useState } from "react";
 
+import { ptBR } from "date-fns/locale/pt-BR";
 import MUIDataTable from "mui-datatables";
-import DateFnsUtils from "@date-io/date-fns";
+import { AdapterDateFns } from "@mui/x-date-pickers/AdapterDateFnsV3";
+import { DatePicker } from "@mui/x-date-pickers/DatePicker";
+import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
 
-import {
-  MuiPickersUtilsProvider,
-  KeyboardDatePicker,
-} from "@material-ui/pickers";
+import Button from "@mui/material/Button";
+import CircularProgress from "@mui/material/CircularProgress";
+import Container from "@mui/material/Container";
+import Dialog from "@mui/material/Dialog";
+import DialogActions from "@mui/material/DialogActions";
+import DialogContent from "@mui/material/DialogContent";
+import DialogTitle from "@mui/material/DialogTitle";
+import FormGroup from "@mui/material/FormGroup";
+import FormLabel from "@mui/material/FormLabel";
+import IconButton from "@mui/material/IconButton";
+import Tooltip from "@mui/material/Tooltip";
 
-import Button from "@material-ui/core/Button";
-import CircularProgress from "@material-ui/core/CircularProgress";
-import Container from "@material-ui/core/Container";
-import Dialog from "@material-ui/core/Dialog";
-import DialogActions from "@material-ui/core/DialogActions";
-import DialogContent from "@material-ui/core/DialogContent";
-import DialogTitle from "@material-ui/core/DialogTitle";
-import FormGroup from "@material-ui/core/FormGroup";
-import FormLabel from "@material-ui/core/FormLabel";
-import IconButton from "@material-ui/core/IconButton";
-import Tooltip from "@material-ui/core/Tooltip";
-
-import DeleteIcon from "@material-ui/icons/Delete";
-import EditIcon from "@material-ui/icons/Edit";
-import PlusOneIcon from "@material-ui/icons/PlusOne";
+import DeleteIcon from "@mui/icons-material/Delete";
+import EditIcon from "@mui/icons-material/Edit";
+import PlusOneIcon from "@mui/icons-material/PlusOne";
 
 import { FormFeedback } from "../FormFeedback";
 import { Loader } from "../Loaders";
@@ -147,7 +145,7 @@ export const TransactionsTable = () => {
     });
 
     Object.entries(multipleChoiceFilters).forEach(([key, value]) =>
-      value.map((v) => _filters.append(key, v))
+      value.map((v) => _filters.append(key, v)),
     );
 
     return _filters.toString();
@@ -209,7 +207,7 @@ export const TransactionsTable = () => {
           getChoiceByLabel(f, [
             ...TransactionsActionsMapping,
             ...AssetsTypesMapping,
-          ]).value
+          ]).value,
       );
       setFilters({ ...filters, [_column]: _filters, page: 1 });
     },
@@ -334,10 +332,13 @@ export const TransactionsTable = () => {
         filterOptions: {
           names: [],
           display: (filterList, onChange, index, column) => (
-            <MuiPickersUtilsProvider utils={DateFnsUtils}>
+            <LocalizationProvider
+              dateAdapter={AdapterDateFns}
+              adapterLocale={ptBR}
+            >
               <FormLabel>Quando</FormLabel>
               <FormGroup row>
-                <KeyboardDatePicker
+                <DatePicker
                   disableToolbar
                   variant="inline"
                   format="dd/MM/yyyy"
@@ -353,7 +354,7 @@ export const TransactionsTable = () => {
                   }}
                   style={{ width: "48%", marginRight: "2%" }}
                 />
-                <KeyboardDatePicker
+                <DatePicker
                   disableToolbar
                   variant="inline"
                   format="dd/MM/yyyy"
@@ -370,7 +371,7 @@ export const TransactionsTable = () => {
                   style={{ width: "48%" }}
                 />
               </FormGroup>
-            </MuiPickersUtilsProvider>
+            </LocalizationProvider>
           ),
         },
       },
