@@ -1,24 +1,22 @@
 import { useEffect, useState } from "react";
 
+import { ptBR } from "date-fns/locale/pt-BR";
 import { useForm, Controller } from "react-hook-form";
-import NumberFormat from "react-number-format";
+import { NumericFormat } from "react-number-format";
 import * as yup from "yup";
 
-import DateFnsUtils from "@date-io/date-fns";
-import {
-  MuiPickersUtilsProvider,
-  KeyboardDatePicker,
-} from "@material-ui/pickers";
+import { AdapterDateFns } from "@mui/x-date-pickers/AdapterDateFnsV3";
+import { DatePicker } from "@mui/x-date-pickers/DatePicker";
+import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
+import Autocomplete from "@mui/lab/Autocomplete";
 
-import Autocomplete from "@material-ui/lab/Autocomplete";
-
-import Button from "@material-ui/core/Button";
-import CircularProgress from "@material-ui/core/CircularProgress";
-import DialogActions from "@material-ui/core/DialogActions";
-import FormControl from "@material-ui/core/FormControl";
-import FormGroup from "@material-ui/core/FormGroup";
-import FormHelperText from "@material-ui/core/FormHelperText";
-import TextField from "@material-ui/core/TextField";
+import Button from "@mui/material/Button";
+import CircularProgress from "@mui/material/CircularProgress";
+import DialogActions from "@mui/material/DialogActions";
+import FormControl from "@mui/material/FormControl";
+import FormGroup from "@mui/material/FormGroup";
+import FormHelperText from "@mui/material/FormHelperText";
+import TextField from "@mui/material/TextField";
 
 import { yupResolver } from "@hookform/resolvers/yup";
 
@@ -34,7 +32,7 @@ function NumberFormatCustom(props) {
   const { inputRef, onChange, ...other } = props;
 
   return (
-    <NumberFormat
+    <NumericFormat
       {...other}
       getInputRef={inputRef}
       onValueChange={(values) =>
@@ -48,7 +46,7 @@ function NumberFormatCustom(props) {
       decimalSeparator=","
       decimalScale={2}
       allowNegative={false}
-      isNumericString
+      valueIsNumericString
     />
   );
 }
@@ -79,7 +77,7 @@ const schema = yup.object().shape({
         .required("O tipo de evento é obrigatório")
         .matches(
           /(PROVISIONED|CREDITED)/,
-          "Apenas 'creditado' e 'provisionado' são tipos de eventos válidos"
+          "Apenas 'creditado' e 'provisionado' são tipos de eventos válidos",
         ),
     })
     .required("O tipo de evento é obrigatório")
@@ -108,22 +106,23 @@ export const PassiveIncomeForm = ({
     getChoiceByLabel(initialData.event_type, PassiveIncomeEventTypesMapping)
       ?.value || "CREDITED";
   const [isCreditedIncome, setIsCreditedIncome] = useState(
-    initialEventType === "CREDITED"
+    initialEventType === "CREDITED",
   );
 
   useEffect(
-    () =>
+    () => {
       new AssetsApi().getMinimalData().then((response) => {
         setCodes(
           response.data.map((asset) => ({
             label: asset.code,
             value: asset.pk,
             currency: asset.currency,
-          }))
+          })),
         );
-      }),
+      });
+    },
     // .catch((error) => {})
-    []
+    [],
   );
 
   const {
@@ -304,7 +303,7 @@ export const PassiveIncomeForm = ({
               defaultValue={
                 getChoiceByLabel(
                   initialData.event_type,
-                  PassiveIncomeEventTypesMapping
+                  PassiveIncomeEventTypesMapping,
                 ) || PassiveIncomeEventTypesMapping[1]
               }
               render={({ field: { onChange, value } }) => (
@@ -339,7 +338,10 @@ export const PassiveIncomeForm = ({
               )}
             />
           </FormControl>
-          <MuiPickersUtilsProvider utils={DateFnsUtils}>
+          <LocalizationProvider
+            dateAdapter={AdapterDateFns}
+            adapterLocale={ptBR}
+          >
             <Controller
               name="operation_date"
               control={control}
@@ -350,7 +352,7 @@ export const PassiveIncomeForm = ({
                   : new Date()
               }
               render={({ field: { onChange, value } }) => (
-                <KeyboardDatePicker
+                <DatePicker
                   onChange={onChange}
                   value={value}
                   label="Quando?"
@@ -368,7 +370,7 @@ export const PassiveIncomeForm = ({
                 />
               )}
             />
-          </MuiPickersUtilsProvider>
+          </LocalizationProvider>
         </FormGroup>
         <FormGroup row style={{ marginTop: "10px" }}>
           <Controller
