@@ -1,27 +1,27 @@
 import { useEffect, useState } from "react";
 
 import { useForm, Controller } from "react-hook-form";
-import NumberFormat from "react-number-format";
+import { NumericFormat } from "react-number-format";
 import * as yup from "yup";
 
-import Autocomplete from "@material-ui/lab/Autocomplete";
-import Button from "@material-ui/core/Button";
-import CircularProgress from "@material-ui/core/CircularProgress";
-import Dialog from "@material-ui/core/Dialog";
-import DialogActions from "@material-ui/core/DialogActions";
-import DialogContent from "@material-ui/core/DialogContent";
-import DialogContentText from "@material-ui/core/DialogContentText";
-import DialogTitle from "@material-ui/core/DialogTitle";
-import FormControl from "@material-ui/core/FormControl";
-import FormHelperText from "@material-ui/core/FormHelperText";
-import FormGroup from "@material-ui/core/FormGroup";
-import Table from "@material-ui/core/Table";
-import TableBody from "@material-ui/core/TableBody";
-import TableCell from "@material-ui/core/TableCell";
-import TableContainer from "@material-ui/core/TableContainer";
-import TableHead from "@material-ui/core/TableHead";
-import TableRow from "@material-ui/core/TableRow";
-import TextField from "@material-ui/core/TextField";
+import Autocomplete from "@mui/lab/Autocomplete";
+import Button from "@mui/material/Button";
+import CircularProgress from "@mui/material/CircularProgress";
+import Dialog from "@mui/material/Dialog";
+import DialogActions from "@mui/material/DialogActions";
+import DialogContent from "@mui/material/DialogContent";
+import DialogContentText from "@mui/material/DialogContentText";
+import DialogTitle from "@mui/material/DialogTitle";
+import FormControl from "@mui/material/FormControl";
+import FormHelperText from "@mui/material/FormHelperText";
+import FormGroup from "@mui/material/FormGroup";
+import Table from "@mui/material/Table";
+import TableBody from "@mui/material/TableBody";
+import TableCell from "@mui/material/TableCell";
+import TableContainer from "@mui/material/TableContainer";
+import TableHead from "@mui/material/TableHead";
+import TableRow from "@mui/material/TableRow";
+import TextField from "@mui/material/TextField";
 import { yupResolver } from "@hookform/resolvers/yup";
 
 import { AssetsApi, AssetTransactionsApi } from "../api";
@@ -44,13 +44,13 @@ const SimulateTransactionResponseDialog = ({
       <DialogContent>
         {formData.quantity ? (
           <DialogContentText>{`${formData.quantity?.toLocaleString(
-            "pt-br"
+            "pt-br",
           )} ativos por ${formData.currency} ${formData.price?.toLocaleString(
             "pt-br",
             {
               minimumFractionDigits: 2,
               maximumFractionDigits: 4,
-            }
+            },
           )} (${formData.currency} ${(
             formData.price * formData.quantity
           )?.toLocaleString("pt-br", {
@@ -68,7 +68,7 @@ const SimulateTransactionResponseDialog = ({
             {
               minimumFractionDigits: 2,
               maximumFractionDigits: 4,
-            }
+            },
           )}  (${(formData.total / formData.price)?.toLocaleString("pt-br", {
             minimumFractionDigits: 2,
             maximumFractionDigits: 4,
@@ -99,7 +99,7 @@ const SimulateTransactionResponseDialog = ({
                     {
                       minimumFractionDigits: 2,
                       maximumFractionDigits: 4,
-                    }
+                    },
                   )}`}
                 </TableCell>
                 <TableCell align="right">
@@ -113,7 +113,7 @@ const SimulateTransactionResponseDialog = ({
                     "pt-br",
                     {
                       minimumFractionDigits: 2,
-                    }
+                    },
                   )}%`}
                 </TableCell>
                 <TableCell align="right">
@@ -122,7 +122,7 @@ const SimulateTransactionResponseDialog = ({
                     {
                       minimumFractionDigits: 2,
                       maximumFractionDigits: 2,
-                    }
+                    },
                   )}`}
                 </TableCell>
               </TableRow>
@@ -138,7 +138,7 @@ const SimulateTransactionResponseDialog = ({
                     {
                       minimumFractionDigits: 2,
                       maximumFractionDigits: 4,
-                    }
+                    },
                   )}`}
                 </TableCell>
                 <TableCell align="right">
@@ -152,7 +152,7 @@ const SimulateTransactionResponseDialog = ({
                     "pt-br",
                     {
                       minimumFractionDigits: 2,
-                    }
+                    },
                   )}%`}
                 </TableCell>
                 <TableCell align="right">
@@ -161,7 +161,7 @@ const SimulateTransactionResponseDialog = ({
                     {
                       minimumFractionDigits: 2,
                       maximumFractionDigits: 2,
-                    }
+                    },
                   )}`}
                 </TableCell>
               </TableRow>
@@ -181,7 +181,7 @@ function NumberFormatCustom(props) {
   const { inputRef, onChange, ...other } = props;
 
   return (
-    <NumberFormat
+    <NumericFormat
       {...other}
       getInputRef={inputRef}
       onValueChange={(values) =>
@@ -195,7 +195,7 @@ function NumberFormatCustom(props) {
       decimalSeparator=","
       decimalScale={8}
       allowNegative={false}
-      isNumericString
+      valueIsNumericString
     />
   );
 }
@@ -215,23 +215,23 @@ const schema = yup.object().shape(
       .required("O preço é obrigatório")
       .positive("Apenas números positivos"),
     quantity: yup.number().when("total", {
-      is: (total) => !total,
-      then: yup
-        .number()
-        .required("Se 'Total' não for inserido, 'Quantidade' é obrigatório")
-        .positive("Apenas números positivos"),
-      otherwise: yup.number().positive("Apenas números positivos"),
+      to: (total) => !total,
+      then: (schema) =>
+        schema
+          .required("Se 'Total' não for inserido, 'Quantidade' é obrigatório")
+          .positive("Apenas números positivos"),
+      otherwise: (schema) => schema.positive("Apenas números positivos"),
     }),
     total: yup.number().when("quantity", {
-      is: (quantity) => !quantity,
-      then: yup
-        .number()
-        .required("Se 'Quantidade' não for inserido, 'Total' é obrigatório")
-        .positive("Apenas números positivos"),
-      otherwise: yup.number().positive("Apenas números positivos"),
+      to: (quantity) => !quantity,
+      then: (schema) =>
+        schema
+          .required("Se 'Quantidade' não for inserido, 'Total' é obrigatório")
+          .positive("Apenas números positivos"),
+      otherwise: (schema) => schema.positive("Apenas números positivos"),
     }),
   },
-  [["total", "quantity"]]
+  [["total", "quantity"]],
 );
 
 export const SimulateTransactionForm = ({ handleClose }) => {
@@ -242,7 +242,7 @@ export const SimulateTransactionForm = ({ handleClose }) => {
   const [formData, setFormData] = useState({});
 
   useEffect(
-    () =>
+    () => {
       new AssetsApi().getMinimalData().then((response) => {
         setCodes(
           response.data.map((asset) => ({
@@ -253,11 +253,12 @@ export const SimulateTransactionForm = ({ handleClose }) => {
                 ? "R$"
                 : "$"
               : "",
-          }))
+          })),
         );
-      }),
+      });
+    },
     // .catch((error) => {})
-    []
+    [],
   );
 
   const {
