@@ -1,27 +1,25 @@
 import { useEffect, useState } from "react";
 
+import { ptBR } from "date-fns/locale/pt-BR";
 import { useForm, Controller } from "react-hook-form";
-import NumberFormat from "react-number-format";
+import { NumericFormat } from "react-number-format";
 import * as yup from "yup";
 
-import DateFnsUtils from "@date-io/date-fns";
-import {
-  MuiPickersUtilsProvider,
-  KeyboardDatePicker,
-} from "@material-ui/pickers";
+import { AdapterDateFns } from "@mui/x-date-pickers/AdapterDateFnsV3";
+import { DatePicker } from "@mui/x-date-pickers/DatePicker";
+import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
+import Autocomplete from "@mui/lab/Autocomplete";
 
-import Autocomplete from "@material-ui/lab/Autocomplete";
-
-import Button from "@material-ui/core/Button";
-import CircularProgress from "@material-ui/core/CircularProgress";
-import DialogActions from "@material-ui/core/DialogActions";
-import FormControl from "@material-ui/core/FormControl";
-import FormControlLabel from "@material-ui/core/FormControlLabel";
-import FormGroup from "@material-ui/core/FormGroup";
-import FormHelperText from "@material-ui/core/FormHelperText";
-import Radio from "@material-ui/core/Radio";
-import RadioGroup from "@material-ui/core/RadioGroup";
-import TextField from "@material-ui/core/TextField";
+import Button from "@mui/material/Button";
+import CircularProgress from "@mui/material/CircularProgress";
+import DialogActions from "@mui/material/DialogActions";
+import FormControl from "@mui/material/FormControl";
+import FormControlLabel from "@mui/material/FormControlLabel";
+import FormGroup from "@mui/material/FormGroup";
+import FormHelperText from "@mui/material/FormHelperText";
+import Radio from "@mui/material/Radio";
+import RadioGroup from "@mui/material/RadioGroup";
+import TextField from "@mui/material/TextField";
 
 import { yupResolver } from "@hookform/resolvers/yup";
 
@@ -34,7 +32,7 @@ function NumberFormatCustom(props) {
   const { inputRef, onChange, ...other } = props;
 
   return (
-    <NumberFormat
+    <NumericFormat
       {...other}
       getInputRef={inputRef}
       onValueChange={(values) =>
@@ -48,7 +46,7 @@ function NumberFormatCustom(props) {
       decimalSeparator=","
       decimalScale={4}
       allowNegative={false}
-      isNumericString
+      valueIsNumericString
     />
   );
 }
@@ -94,18 +92,19 @@ export const TransactionForm = ({ initialData, handleClose, reloadTable }) => {
   let api = new AssetsApi();
 
   useEffect(
-    () =>
+    () => {
       api.getMinimalData().then((response) => {
         setCodes(
           response.data.map((asset) => ({
             label: asset.code,
             value: asset.pk,
             currency: asset.currency,
-          }))
+          })),
         );
-      }),
+      });
+    },
     // .catch((error) => {})
-    []
+    [],
   );
 
   const {
@@ -282,7 +281,10 @@ export const TransactionForm = ({ initialData, handleClose, reloadTable }) => {
               />
             )}
           />
-          <MuiPickersUtilsProvider utils={DateFnsUtils}>
+          <LocalizationProvider
+            dateAdapter={AdapterDateFns}
+            adapterLocale={ptBR}
+          >
             <Controller
               name="operation_date"
               control={control}
@@ -293,7 +295,7 @@ export const TransactionForm = ({ initialData, handleClose, reloadTable }) => {
                   : new Date()
               }
               render={({ field: { onChange, value } }) => (
-                <KeyboardDatePicker
+                <DatePicker
                   onChange={onChange}
                   value={value}
                   label="Quando?"
@@ -311,7 +313,7 @@ export const TransactionForm = ({ initialData, handleClose, reloadTable }) => {
                 />
               )}
             />
-          </MuiPickersUtilsProvider>
+          </LocalizationProvider>
         </FormGroup>
         <FormGroup row style={{ marginTop: "10px" }}>
           <Controller
@@ -337,7 +339,7 @@ export const TransactionForm = ({ initialData, handleClose, reloadTable }) => {
                 error={!!errors.current_currency_conversion_rate}
                 helperText={
                   errors.current_currency_conversion_rate?.message ||
-                  `O valor de 1 ${assetData?.currency}, em reais, no dia que a operação de venda foi realizada`
+                  `O valor de 1 ${assetData?.currency}, em reais, no dia que a operação foi realizada`
                 }
               />
             )}
